# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "Guardian"
  description: "The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation."
  version: "2.9.3"
  contact:
    name: "API developer"
    url: "https://envisionblockchain.com"
    email: "info@envisionblockchain.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
- url: /api/v1
  description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
...
    User:
      type: object
      required:
        - username
        - role
      properties:
        confirmed:
          type: string
        failed:
          type: string
        username:
          type: string
        role:
          type: string
        hederaAccountId:
          type: string
        hederaAccountKey:
          type: string
        did:
          type: string
        didDocument:
            type: object
        vcDocument:
            type: object
        parent:
          type: string
        topicId:
          type: string
    Schema:
      type: object
      required:
        - id
        - name
        - description
        - entity
        - document
      properties:
        id:
          type: string
        iri:
          type: string
        uuid:
          type: string
        name:
          type: string
        description:
          type: string
        entity:
          type: string
        hash:
          type: string
        status:
          type: string
        document:
          type: string
        topicId:
          type: string
        version:
          type: string
        owner:
          type: string
        messageId:
          type: string
...
    PolicyConfig:
      type: object
      required:
        - name
        - version
        - description
        - topicDescription
        - config
        - topicId
        - policyTag
      properties:
        id:
          type: string
        uuid:
          type: string
        name:
          type: string
        version:
          type: string
        description:
          type: string
        topicDescription:
          type: string
        config:
          type: object
        status:
          type: string
        owner:
          type: string
        policyRoles:
          type: array
          items:
            type: string
        topicId:
          type: string
        policyTag:
          type: string
        policyTopics:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              type:
                type: string
              static:
                type: boolean
...
    ExportPolicy:
      type: object
      required:
        - name
        - version
        - messageId
      properties:
        name:
          type: string
        version:
          type: string
        tokens:
          type: string
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    Task:
      type: object
      properties:
        taskId:
          type: string
        expectation:
          type: number
...
paths:
...
  /profiles/{username}/balance:
    get:
      tags:
        - profiles
      description: Requests Hedera account balance. Only users with the Installer role are allowed to make the request.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user for whom to fetch the balance.
      security:
      - bearerAuth: []
      summary: Returns user's Hedera account balance.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: string
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /profiles/{username}:
    get:
      tags:
        - profiles
      description: Returns user account information. For users with the Standard Registry role it also returns address book and VC document information.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user for whom to fetch the information.
      security:
      - bearerAuth: []
      summary: Returns user account info.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - profiles
      description: Sets Hedera credentials for the user. For users with the Standard Registry role it also creates an address book.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user for whom to update the information.
      security:
      - bearerAuth: []
      summary: Sets Hedera credentials for the user.
      requestBody:
        description: Object that contains the Hedera account data.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        201:
          description: Successful operation.
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /schemas/{topicId}/import/file:
    post:
      tags:
        - schemas
      description: Imports new schema from a zip file into the local DB. Only users with the Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: topicId
          schema:
            type: string
          required: true
          description: Topic ID.
          examples:
            topicId:
              summary: Example of a Topic ID
              value: 0.0.00000001
      security:
      - bearerAuth: []
      summary: Imports new schema from a zip file.
      requestBody:
        description: A zip file containing schema to be imported.
        content:
          binary/octet-stream:
            schema:
              type: string
              format: binary
        required: true
      responses:
        201:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schemas/push/{topicId}/import/file:
    post:
      tags:
        - schemas
      description: Imports new schema from a zip file into the local DB. Only users with the Standard Registry role are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Imports new schema from a zip file.
      requestBody:
        description: A zip file containing schema to be imported.
        content:
          binary/octet-stream:
            schema:
              type: string
              format: binary
        required: true
      parameters:
        - in: path
          name: topicId
          schema:
            type: string
          required: true
          description: Topic ID.
          examples:
            topicId:
              summary: Example of a Topic ID
              value: 0.0.00000001
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /schemas/{schemaId}/export/file:
    post:
      tags:
        - schemas
      description: Returns schema files for the schemas. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Return zip file with schemas.
      parameters:
        - in: path
          name: schemaId
          schema:
            type: string
          required: true
          description: Selected schema ID.
      responses:
        200:
          description: Successful operation. Response zip file
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /policies/{policyId}/export/file:
    get:
      tags:
        - policies
      description: Returns a zip file containing the published policy and all associated artifacts, i.e. schemas and VCs. Only users with the Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: policyId
          schema:
            type: string
          required: true
          description: Selected policy ID.
      security:
      - bearerAuth: []
      summary: Return policy and its artifacts in a zip file format for the specified policy.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPolicy'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /policies/import/file:
    post:
      tags:
        - policies
      description: Imports new policy and all associated artifacts, such as schemas and VCs, from the provided zip file into the local DB. Only users with the Standard Registry role are allowed to make the request.
      parameters:
        - in: query
          name: versionOfTopicId
          schema:
            type: string
          description: The topic ID of policy version.
          examples:
            pageIndex:
              summary: Example of a topic ID of policy version.
              value: '0.0.00000001'
      security:
      - bearerAuth: []
      summary: Imports new policy from a zip file.
      requestBody:
        description: A zip file that contains the policy and associated schemas and VCs to be imported.
        required: true
        content:
          binary/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        201:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PolicyConfig'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /policies/push/import/file:
    post:
      tags:
        - policies
      description: Imports new policy and all associated artifacts, such as schemas and VCs, from the provided zip file into the local DB. Only users with the Standard Registry role are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Imports new policy and all associated artifacts, such as schemas and VCs, from the provided zip file into the local DB.
      requestBody:
        description: A zip file that contains the policy and associated schemas and VCs to be imported.
        required: true
        content:
          binary/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /ipfs/file/{cid}:
    get:
      tags:
        - ipfs
      description: Get file from ipfs.
      summary: Get file from ipfs.
      parameters:
        - in: path
          name: cid
          schema:
            type: string
          required: true
          description: File CID.
      security:
      - bearerAuth: []
      responses:
        201:
          description: Successful operation.
          content:
            binary/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ipfs/file/:
    post:
      tags:
        - ipfs
      description: Add file to ipfs.
      summary: Add file to ipfs.
      requestBody:
        description: Data array of file.
        required: true
        content:
          binary/octet-stream:
              schema:
                type: string
                format: binary
      security:
      - bearerAuth: []
      responses:
        201:
          description: Successful operation.
          content:
            application/json:
              schema:
                description: CID of added file.
                type: string
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...